backend/requirements.txt
Flask
boto3
transformers
torch
python-dotenv